fix(ci): format, split mcp serve parse test under 1000 lines
cli.rs was 1028 lines (cap 1000). Move the AOS plugin argv parse test
out and rustfmt the MCP/audit-sink edits.
