+++ crates/astrid-cli/src/cli.rs
@@ -489,15 +489,10 @@ pub(crate) enum McpCommands {
     /// stream (EOF) or the process is killed. Stdout carries the MCP
     /// JSON-RPC protocol only — all diagnostics go to stderr.
     Serve {
-        /// Project directory this MCP session attaches to.
-        ///
-        /// AOS host plugins pass the client's `$PWD` here. Distinct from
-        /// `--workspace-state-dir`, which names the per-project state folder.
+        /// Project directory AOS host plugins pass as `$PWD`.
         #[arg(long, value_name = "PATH")]
         workspace: Option<PathBuf>,
-        /// Host-plugin compatibility flag. Tool deadlines are owned by the MCP
-        /// client (`tool_timeout_sec`); this value is accepted and ignored so
-        /// `aos mcp serve --request-timeout 1d5m` does not fail clap.
+        /// Accepted and ignored; MCP hosts own `tool_timeout_sec`.
         #[arg(long = "request-timeout", value_name = "DURATION")]
         request_timeout: Option<String>,
     },
@@ -610,6 +605,10 @@ pub(crate) enum DistroCommands {
 mod distro_tests;
 
 #[cfg(test)]
+#[path = "cli_mcp_tests.rs"]
+mod mcp_tests;
+
+#[cfg(test)]
 #[path = "cli_capsule_tests.rs"]
 mod capsule_tests;
 
@@ -635,35 +634,6 @@ mod tests {
             alternate.workspace_state_dir.state_dir_name(),
             ".alternate-runtime"
         );
-    }
-
-    #[test]
-    fn mcp_serve_accepts_aos_host_plugin_flags() {
-        let parsed = Cli::try_parse_from([
-            "astrid",
-            "--principal",
-            "codex-code",
-            "mcp",
-            "serve",
-            "--workspace",
-            "/tmp/project",
-            "--request-timeout",
-            "1d5m",
-        ])
-        .expect("AOS plugin argv must parse");
-        match parsed.command {
-            Some(Commands::Mcp {
-                command:
-                    McpCommands::Serve {
-                        workspace,
-                        request_timeout,
-                    },
-            }) => {
-                assert_eq!(workspace.as_deref(), Some(std::path::Path::new("/tmp/project")));
-                assert_eq!(request_timeout.as_deref(), Some("1d5m"));
-            },
-            _ => panic!("expected mcp serve"),
-        }
     }
 
     #[test]
